data validation logic added

+++ Controllers/CommentController.cs
@@ -23,6 +23,9 @@ namespace MovieMania.Controllers
         [HttpGet]
         public async Task<IActionResult> GetAll()
         {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+
             var comments = await _commentRepository.GetAllAsync();
             var commentDto = comments.Select(s => s.ToCommentDto());
             return Ok(commentDto);
@@ -30,9 +33,12 @@ namespace MovieMania.Controllers
 
 
         [HttpGet]
-        [Route("{id}")]
+        [Route("{id:int}")]
             public async Task<IActionResult> GetById([FromRoute] int id)
         {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+
             var comment = await _commentRepository.GetByIdAsync(id);
 
             if(comment == null)
@@ -44,9 +50,11 @@ namespace MovieMania.Controllers
         }
 
         [HttpPost]
-        [Route("{movieId}")]
+        [Route("{movieId:int}")]
         public async Task<IActionResult> Create( [FromRoute] int movieId, CreateCommentDto commentDto )
         {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
             if (!await _movieRepository.MovieExists(movieId))
             {
                 return BadRequest("Movie does not exist!!");
@@ -57,9 +65,11 @@ namespace MovieMania.Controllers
         }
 
         [HttpDelete]
-        [Route("{id")]
+        [Route("{id:int}")]
         public async Task<IActionResult> Delete([FromRoute] int id)
         {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
             var commentModel = await _commentRepository.DeleteAsync(id);
